requirements.txt
Django==1.9.1
dj-database-url==0.3.0
gunicorn==19.1.1
psycopg2==2.5.1
python-social-auth
django-adminrestrict
django-csp==2.0.3
Fabric==1.7.0
gnacs
-e git://github.com/DrSkippy/Simple-n-grams.git@bbfd782614b39e2d0a1bc01fc6a75cc5df235e3e#egg=Simple-n-grams
